requirements.txt
aiodns==3.0.0
aioflask==0.4.0
aiosmtpd==1.4.2
asgiref==3.5.2
atpublic==3.0.1
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
dnspython==2.2.1
Flask==2.1.2
greenlet==1.1.2
greenletio==0.9.0
h11==0.13.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
nest-asyncio==1.5.5
numpy==1.22.4
pandas==1.4.2
pendulum==2.1.2
pycares==4.1.2
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.1
pytzdata==2020.1
requests==2.28.0
six==1.16.0
urllib3==1.26.9
uvicorn==0.17.6
verify-email==2.4.3
Werkzeug==2.1.2
zipp==3.8.0
